pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "Gulaschprogrammiernacht Chat"
version = "0.1.0"
description = "A GPT that is trained on the Gulaschprogrammiernacht Talks"
authors = ["Felix Schneider <mail@fschneider.me>", "Vincent Mielke <mielke.vincent@gmail.com>"]
package-mode = false


[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<4.0"
streamlit = "^1.38.0"
haystack-ai = "^2.5.1"
bs4 = "^0.0.2"
python-dotenv = "^1.0.1"
openai-whisper = {git = "https://github.com/openai/whisper.git"}
black = "^24.8.0"
isort = "^5.13.2"
transformers = {extras = ["rag"], version = "^4.44.2"}
sentence-transformers = "^3.1.0"
qdrant-client = "^1.11.1"
qdrant-haystack = "^6.0.0"
gitpython = "^3.1.43"
docker = "^7.1.0"
ollama-haystack = "^0.0.7"
pydub = "^0.25.1"
torch = {version = "^2.4.1+cu118", source = "pytorch-gpu"}
torchvision = {version = "^0.19.1+cu118", source = "pytorch-gpu"}
torchaudio = {version = "^2.4.1+cu118", source = "pytorch-gpu"}

[[tool.poetry.source]]
name = "pytorch-gpu"
url = "https://download.pytorch.org/whl/cu118"
priority = "explicit"
